Move Drawer and Footer components under common folder
